Variant calling for ancestral genomes
breseq was run with polymorphism mode to identify any mutations already present in the ancestral genomes. The original strain (TDA1000) was sequenced with Illumina MiSeq, whereas the fluorescently-labelled ancestral strains (TDA1001 and TDA1002) were sequenced with both MiSeq and NextSeq, the latter of which was used for sequencing the evolved populations. This makes a total of 5 sequencing runs, all of which were run in breseq to detect mutations already present in the ancestral.